lead-ai/crm/backend/requirements.txt
# ============================================================
# IBMP CRM — Backend Dependencies
# Deduped & pinned — May 2026
# Run: pip install -r requirements.txt --break-system-packages
# ============================================================

# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-multipart==0.0.12

# ── Data Validation ───────────────────────────────────────
pydantic==2.10.3
email-validator==2.2.0

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
supabase==2.10.0
postgrest==0.18.0

# ── Auth & Security ───────────────────────────────────────
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# ── ML / AI ───────────────────────────────────────────────
catboost==1.2.7
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2
openai==1.57.0

# ── Communication Integrations ────────────────────────────
twilio==9.3.7
resend==2.5.0
sendgrid==6.11.0
jinja2==3.1.4

# ── Google Integrations ───────────────────────────────────
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0

# ── HTTP Clients ──────────────────────────────────────────
httpx==0.27.2
aiohttp==3.11.10
requests==2.32.3
tenacity==9.0.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.2
aiofiles==24.1.0
cachetools==5.5.0
apscheduler==3.10.4
loguru==0.7.3

# ── Monitoring & Observability ────────────────────────────
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.19.0

# ── Testing ───────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# ── Linting / Formatting ──────────────────────────────────
ruff==0.8.4
